\name{CodeExtractoR-package}
\alias{CodeExtractoR-package}
\alias{CodeExtractoR}
\docType{package}
\title{\packageTitle{CodeExtractoR}}
\description{\packageDescription{CodeExtractoR}}
\section{Package Content}{\packageIndices{CodeExtractoR}}
\author{\packageAuthor{CodeExtractoR}}
\section{Maintainer}{\packageMaintainer{CodeExtractoR}}
\keyword{package}
