% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{convert_pdf_2_html}
\alias{convert_pdf_2_html}
\title{Convert PDF to HTML}
\usage{
convert_pdf_2_html(
  input_file_url,
  output_file = NULL,
  quiet = FALSE,
  api_key = NULL
)
}
\arguments{
\item{input_file_url}{A \strong{URL} to pdf file to process. Do \strong{not} supply path to file in your filesystem.}

\item{output_file}{A name of file to write .html to. Must be of .html extension. By default, the name of the file is extracted from \code{input_file_url}.}

\item{quiet}{Boolean. Should the method display messages about progress? Errors will be displayed regardless of this option.}

\item{api_key}{A custom API key obtained from \href{https://cloudconvert.com/pdf-to-html}{cloudconvert}. Make sure to use the key of \strong{version 1}.}
}
\value{
A HTML file with name \code{output_file} is created. It is ready to be passed into \code{\link{extract_code_from_html}} function.
}
\description{
This is wrapper around API to \href{https://cloudconvert.com/pdf-to-html}{cloudconvert}.
}
\details{
Create your own account at \href{https://cloudconvert.com/register}{cloudconvert} and obtain your own API key, or convert the file(s) manually.
}
\seealso{
\code{\link{extract_code_from_html}}, \code{\link{extract_code_from_pdf}}
}
