% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiv2.R
\name{convert_pdf_2_html_v2}
\alias{convert_pdf_2_html_v2}
\title{Convert PDF to HTML}
\usage{
convert_pdf_2_html_v2(input_file, output_file = NULL)
}
\description{
This is wrapper around API to \href{https://cloudconvert.com/pdf-to-html}{cloudconvert}.
}
