% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_html.R
\name{extract_code_from_html}
\alias{extract_code_from_html}
\title{Extract code snippets from HTML file}
\usage{
extract_code_from_html(input_file, output_file)
}
\arguments{
\item{input_file}{name of file to extract code from. Must be of HTML extension \strong{and} be produced by converter using pdf2htmlex (like \href{https://cloudconvert.com/pdf-to-html}{cloudconvert}).}

\item{output_file}{name of file to write code to. Must be of R extension.}
}
\description{
Search for code snippets, based on their unique font. Some extra may be extracted (like in-line mentions of functions or HTTP links).
}
