% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_code_from_pdf}
\alias{extract_code_from_pdf}
\title{Extract code from PDF file}
\usage{
extract_code_from_pdf(
  input_file_url,
  output_file,
  output_html_file = NULL,
  quiet = FALSE,
  api_key = NULL,
  clear = FALSE
)
}
\arguments{
\item{input_file_url}{A \strong{URL} to pdf file to process. Do \strong{not} supply path to file in your filesystem.}

\item{output_file}{name of file to write code to. Must be of R extension.}

\item{output_html_file}{name of HTML file created inside function.
\itemize{
\item{if \code{NULL}, then the name of the file is extracted from the \code{input_file_url} and the file is created in working directory.
      if it exists and \code{clear} is \code{FALSE} then it is assumed to be \strong{product of convertion}, that already took place, and code is extracted from it.}
\item{if supplied and a file of that name exists, it is assumed to be a \strong{mistake}.}     
}}

\item{quiet}{Boolean. Should the method display messages about progress? Errors will be displayed regardless of this option.}

\item{api_key}{A custom API key obtained from \href{https://cloudconvert.com/pdf-to-html}{cloudconvert}. Make sure to use the key of \strong{version 1}.}

\item{clear}{Should the HTML file be removed? If file exists before conversion, it will be removed too - even when \code{output_html_file} was not supplied.}
}
\description{
Extract code content from PDF file and write it to .R file.
}
