% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_html.R
\name{extract_code_from_html}
\alias{extract_code_from_html}
\title{Extract code snippets from HTML file}
\usage{
extract_code_from_html(
  input_file,
  output_file = NULL,
  filter = TRUE,
  bibliography = FALSE,
  console_char = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{input_file}{name of file to extract code from. Must be of HTML extension \strong{and} be produced by converter using pdf2htmlex (like \href{https://cloudconvert.com/pdf-to-html}{cloudconvert}).}

\item{output_file}{name of file to write code to. Must be of R extension. By default its name is extracted from \code{input_file}.}

\item{filter}{Boolean. Should all words written with \code{code font} be extracted, or should they be filted first? See Details.}

\item{bibliography}{Boolean. Should the words be extracted from bibliography too?}

\item{console_char}{Single character used in document to indicate start of a command. Assumed to be regex. If supported, from all lines starting with this character the character will be removed and all lines \strong{not} starting with this character will be commented.}

\item{overwrite}{Boolean. Should the output file (.R) be overwritten?}
}
\description{
Search for code snippets, based on their unique font. Some extra may be extracted (like in-line mentions of functions or HTTP links).
}
\details{
Often the unfiltered results contain unnecessary occurrences of words written with \code{code} font. 
If \code{filter} is set to \code{TRUE}, it is assumed, that we are only interested in words, that are \strong{not} accompanied by words in different fonts in the same line.
This prevents us from extracting names of functions mentioned inside paragraphs of ordinary text, for example.
}
\seealso{
\code{\link{extract_code_from_pdf}}
}
