% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_code_from_pdf.R
\name{extract_code_from_pdf}
\alias{extract_code_from_pdf}
\title{Extract code from PDF file}
\usage{
extract_code_from_pdf(
  input_file_url,
  output_file = NULL,
  output_html_file = NULL,
  filter = TRUE,
  bibliography = FALSE,
  console_char = NULL,
  quiet = FALSE,
  api_key = NULL,
  clear = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{input_file_url}{A \strong{URL} to pdf file to process. Do \strong{not} supply path to file in your filesystem.}

\item{output_file}{name of file to write code to. Must be of R extension. By default its name is extracted from \code{input_file_url}.}

\item{output_html_file}{name of HTML file created inside function.
\itemize{
\item{if \code{NULL}, then the name of the file is extracted from the \code{input_file_url} and the file is created in working directory.
      if it exists and \code{clear} is \code{FALSE} then it is assumed to be \strong{product of conversion}, that already took place, and code is extracted from it.}
\item{if supplied and a file of that name exists, it is assumed to be a \strong{mistake}.}     
}}

\item{filter}{Boolean. Should all words written with \code{code font} be extracted, or should they be filtered first? See Details.}

\item{bibliography}{Boolean. Should the words be extracted from bibliography too?}

\item{console_char}{Single character used in document to indicate start of a command. Assumed to be regex. If supported, from all lines starting with this character the character will be removed and all lines \strong{not} starting with this character will be commented.}

\item{quiet}{Boolean. Should the method display messages about progress? Errors will be displayed regardless of this option.}

\item{api_key}{A custom API key obtained from \href{https://cloudconvert.com/pdf-to-html}{cloudconvert}. Make sure to use the key of \strong{version 1}.}

\item{clear}{Should the HTML file be removed? If file exists before conversion, it will be removed too - even when \code{output_html_file} was not supplied.}

\item{overwrite}{Boolean - should the files created throughout the function (.R and .html) be overwritten?}
}
\value{
An .R file is created, containing text \strong{suspected} to be code. Unfortunately, various articles use various fonts for the code included.
The texts in .R file are grouped based on their font in the \code{input_file}. 
User is encouraged to compare the results with the article.
}
\description{
Extract code content from PDF file and write it to .R file.
}
\details{
Often the unfiltered results contain unnecessary occurrences of words written with \code{code} font.
If \code{filter} is set to \code{TRUE}, it is assumed, that we are only interested in words, that are \strong{not} accompanied by words in different fonts in the same line.
This prevents us from extracting names of functions mentioned inside paragraphs of ordinary text, for example.
}
\seealso{
\code{\link{extract_code_from_html}}, \code{\link{convert_pdf_2_html}}
}
